todo server 작업 중

+++ src/main/java/com/kcnet/todosv/card/Card.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 public class Card {
   private String cardId;
   private String description;
-  private int position;
+  private double position;
 
 
 }
